Add tests for PathArc::absolute() method.
Tests are added as integration tests, because Cargo likes to run tests in
parallel in multiple threads, while "the process current working directory" is a
mutable global variable. If we put all the tests that require a particular value
for the current directory into a separate process, they won't stomp on each
other when run in parallel.

Also, adds portability comments to tests I have observed to fail.