Fix WASI symlink implementation
As described in #51, current implementation is incorrect and won't work and, as described in https://github.com/rust-lang/rust/issues/68574, there was no correct way to use that API.

I went ahead and added a new API for symlinks in https://github.com/rust-lang/rust/pull/81542, and now that it's landed, it can be used to correctly create symlinks.

Fixes #51.